requirements.txt
astroid==3.3.8
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
Flask==3.1.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.3
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
tomlkit==0.13.2
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
